+++ ConsoleAPI.h
@@ -1,0 +1,9 @@
+#ifndef CONSOLEAPI_H_INCLUDED
+#define CONSOLEAPI_H_INCLUDED
+
+void goToXY(int x, int y);
+void ShowConsoleCursor(bool showFlag);
+void SetTextColor(int ForgC);
+void ConsoleSetup();
+
+#endif // CONSOLEAPI_H_INCLUDED
