+++ Declaration.h
@@ -1,0 +1,45 @@
+#ifndef DECLARATION_H_INCLUDED
+#define DECLARATION_H_INCLUDED
+#include <string>
+
+// CONSTANTS
+
+#define Debug   false
+
+#define KEY_UP      72
+#define KEY_DOWN    80
+#define KEY_LEFT    75
+#define KEY_RIGHT   77
+
+#define foodIcon        254
+#define snakeHeadIcon   178
+#define snakeTailIcon   176
+
+#define defaultColor     15
+#define snakeHeadColor   2
+#define snakeTailColor   14
+
+#define MAX_TAIL (W-2)*(H-2)
+
+/*
+ +Name         | Value
+ +             |
+ +Black        |   0
+ +Blue         |   1
+ +Green        |   2
+ +Cyan         |   3
+ +Red          |   4
+ +Magenta      |   5
+ +Brown        |   6
+ +Light Gray   |   7
+ +Dark Gray    |   8
+ +Light Blue   |   9
+ +Light Green  |   10
+ +Light Cyan   |   11
+ +Light Red    |   12
+ +Light Magenta|   13
+ +Yellow       |   14
+ +White        |   15
+ +*/
+
+#endif // DECLARATION_H_INCLUDED
