+++ Screens.h
@@ -1,0 +1,26 @@
+#ifndef SCREENS_H_INCLUDED
+#define SCREENS_H_INCLUDED
+
+/*
+    //////////////////////////////////
+    ////           LOGO           ////
+    //////////////////////////////////
+               > START GAME
+                MAP SELECT
+                  ABOUT
+               QUIT GAME :(
+*/
+
+void mainMenu();
+
+void playScreen();
+
+void mapScreen();
+
+void aboutScreen();
+
+void resultScreen();
+
+
+
+#endif // SCREENS_H_INCLUDED
